requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
bitarray==3.8.0
certifi==2026.2.25
charset-normalizer==3.4.4
ckzg==2.1.5
cytoolz==1.1.0
Django==4.2.27
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
hexbytes==1.3.1
idna==3.11
parsimonious==0.10.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
sqlparse==0.5.5
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
chromadb
openai
apscheduler
